Рабочий бэк(кроме ошибок при удалении и изменении шаблона). Надо затестить все эндпоинты. Подкоректировать вывод ошибок. Добавить сортировки в контроль. Добавить пагинацию

+++ Back/InvestmentControl/InvestmentControl.Application/Control/Commands/DeleteCostCommand.cs
@@ -4,6 +4,8 @@ using InvestmentControl.Domain.Interfaces;
 using InvestmentControl.Domain.Models;
 using MediatR;
 
+namespace InvestmentControl.Application.Control.Commands;
+
 public class DeleteCostCommand : IRequest
 {
     public int Id { get; set; }
@@ -12,11 +14,13 @@ public class DeleteCostCommand : IRequest
 public class DeleteCostCommandHandler : IRequestHandler<DeleteCostCommand>
 {
     private readonly ICostRepository _costRepository;
+    private readonly IProjectReadRepository _projectReadRepository;
     private readonly ICurrentUser _currentUser;
 
-    public DeleteCostCommandHandler(ICostRepository costRepository, ICurrentUser currentUser)
+    public DeleteCostCommandHandler(ICostRepository costRepository, IProjectReadRepository projectReadRepository, ICurrentUser currentUser)
     {
         _costRepository = costRepository;
+        _projectReadRepository = projectReadRepository;
         _currentUser = currentUser;
     }
 
@@ -28,6 +32,10 @@ public class DeleteCostCommandHandler : IRequestHandler<DeleteCostCommand>
         var cost = await _costRepository.GetByIdAsync(request.Id, cancellationToken);
         if (cost == null)
             throw new NotFoundException(nameof(Cost), request.Id);
+
+        var creatorId = await _projectReadRepository.GetCreatorUserIdAsync(cost.ProjectId, cancellationToken);
+        if (creatorId != _currentUser.UserId)
+            throw new ForbiddenAccessException("Вы можете удалять затраты только для своих проектов.");
 
         _costRepository.Delete(cost);
         await _costRepository.SaveChangesAsync(cancellationToken);
